fix: due to comments.p6

+++ stat-app/stat-server/src/main/java/ru/practicum/stats/service/StatService.java
@@ -11,5 +11,5 @@ public interface StatService {
 
     List<ViewStats> get(Date start, Date end, List<String> uris, Boolean unique);
 
-    Integer getCount(String uri);
+    Long getCount(String uri);
 }
